+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### Unreleased
+* Refactor parameter ordering in `find()` and `draft.send()` methods [backwards compatible]
 * Add JobStatus model and collection
 
 ### 5.2.0 / 2020-07-27